requirements.txt
sanic
django-environ
tortoise-orm
aiomysql
aerich